tests: Move testUtils outside of __tests__ folder
