+++ labware-library/src/labware-creator/components/__tests__/sections/GridOffset.test.tsx
@@ -3,7 +3,7 @@ import { FormikConfig } from 'formik'
 import isEqual from 'lodash/isEqual'
 import { when } from 'jest-when'
 import { render, screen } from '@testing-library/react'
-import { nestedTextMatcher } from '../testUtils'
+import { nestedTextMatcher } from '../../__testUtils__/nestedTextMatcher'
 import { getDefaultFormState, LabwareFields } from '../../../fields'
 import { isEveryFieldHidden } from '../../../utils'
 import { GridOffset } from '../../sections/GridOffset'
